Move main to mono source set
