+++ app/android/src/main/kotlin/activity/AniComponentActivity.kt
@@ -1,5 +1,5 @@
 package me.him188.ani.android.activity
 
-import me.him188.ani.app.activity.BaseComponentActivity
+import me.him188.ani.app.platform.BaseComponentActivity
 
 open class AniComponentActivity : BaseComponentActivity()
