+++ app/shared/src/androidMain/kotlin/platform/BaseComponentActivity.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.him188.ani.app.activity
+package me.him188.ani.app.platform
 
 import androidx.activity.ComponentActivity
 import androidx.activity.SystemBarStyle
