+++ app/shared/src/desktopMain/kotlin/platform/CommonKoinModule.desktop.kt
